Fully working server with x_based client.
Moved the settings into a data structure for tidyness.
Added -x/--xtended_features flag for allowing the usage of x_based
client.

The server is now capable of transmitting and receiving the x_based data
structure that contains the modifier, press/release, and the key.

The server will apply the modifiers to the keys pressed based on the
above data, according to the mapping. Mappings had to be done for
special keys such as pgup, pgdown, home, and punctuation keys.